add app_monitor

+++ core/lib/dal/migrations/20240112140218_application_monitor_add.up.sql
@@ -2,8 +2,8 @@ CREATE TABLE application_monitor (
     id SERIAL NOT NULL PRIMARY KEY ,
     app_name text NOT NULL,
     ip text NOT NULL,
-    start_at INT  NOT NULL,
-    heartbeat_update_at INT  NOT NULL,
+    start_at bigint  NOT NULL,
+    heartbeat_update_at bigint  NOT NULL,
     created_at timestamp(6) without time zone NOT NULL,
     updated_at timestamp(6) without time zone NOT NULL
 );
